pyproject.toml
[project]
name = "wowkmang"
version = "0.1.0"
description = "a claude code orchestrator"
requires-python = ">=3.13"
dependencies = [
    "PyYAML>=6.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.28",
    "docker>=7.0",
    "anthropic>=0.42",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "black>=24.0",
    "mdformat>=0.7",
    "yamlfix>=1.16.0",
    "pre-commit>=3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
